feat(async-api): Add cancellation token for TreeWalker and RuleRunner (#755)
#### Details

Adding cancellation logic to two of the most time intensive operations
performed during scanning, the tree walker and rule runner.

Based on docs/ discussion online, it appears that the norm is to use
`ThrowIfCancellationRequested` to check the cancellation token
throughout the async call. This throws a `OperationCanceledException`,
which will be wrapped in an `AggregateException` once these methods are
wrapped in a task.

Also added bare minimum test coverage for cancellation in
TreeWalkerForTest, which previously did not have any unit tests.

##### Motivation

Part of
[1986365](https://mseng.visualstudio.com/1ES/_workitems/edit/1986365) to
enable cancellation

##### Context

This is the first in what will be multiple PRs to add cancellation
tokens. In the future we will need to add the logic to pass the
cancellation token to the methods changed in this PR, but that is
blocked until we add a top level async scan API. .

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Contributes to
[1986365](https://mseng.visualstudio.com/1ES/_workitems/edit/1986365)