feat: Allow callers to override default DPI awareness handling (#751)
#### Details

We need to be DPI-aware when running an Axe.Windows scan. We currently
call SetProcessDPIAware then never reset the state, partially because
there's no way to obtain the original state and partially because we
never needed to revert to a non-DPI aware state. #733 asks for a way to
scope this and/or reset the original state. We want to enable this, but
since different scenarios could require different handling, we are
moving the DPI awareness into an interface and allowing the caller to
provide their own implementation. The new interface is defined as
follows:

```
    public interface IDPIAwareness
    {
        /// <summary>
        /// Enable DPI awareness for the scan
        /// </summary>
        /// <returns>An object that will be passed as a parameter <see cref="IDPIAwareness.Restore"/></returns>
        object Enable();

        /// <summary>
        /// Restore DPI awareness to its original state
        /// </summary>
        /// <param name="dataFromEnable">The object returned from the call to <see cref="IDPIAwareness.Enable"/></param>
        void Restore(object dataFromEnable);
    }
```

Calls to `Enable` and `Restore` will always be paired. When we add async
support, we'll ensure that both calls will occur on the thread where the
scanning will occur. We'll need to work out the details for
ScanAllAsync, but if we set this up to have the scans run concurrently,
then a thread will be created for each scan and each scan will have a
pair of `Enable`/`Restore` calls.

A user's implementation of `IDPIAwareness` is injected by calling
`Config.Builder.WithDPIAwareness` when setting up the scanner. Apps that
don't call `ConfigBuilder.WithDPIAwareness` will continue to use the
existing behavior, so backward compatibility is preserved.

##### Motivation

Address #733 

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->
We _could_ have just set the DPI awareness on the thread, then reset it
as suggested in a comment on #733. That would potentially break backward
compatibility, and it opens the door of possibly trying to provide
one-off solutions as customer needs change. The simpler long-term option
is to allow customers to implement what works for their scenario. This
may become even more important as we think about adding async versions
of the scanning methods. Making assumptions about the calling code is a
great way to create a bug pump.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000