feat(async-api): Provide separate contexts for each scan (#757)
#### Details

I apologize in advance for the size of this PR. I wasn't able to
identify good ways to split it up into smaller chunks ☹️

The async work requires us to support concurrent scans. The existing
code can't handle this, as all data lives in a big singleton that is
accessed from multiple classes. This creates an `IActionContext`
abstraction that is intended to provide a separate data space for each
action. The `IActionContext` includes the following:
Field | Content
--- | ---
`DataManager` | The `DataManager` object that stores the UIA hierarchy
and associated properties
`SelectAction` | A `SelectAction` object that is bound to the underlying
`DataManager` object
`Registrar` | A `Registrar` object for managing custom UIA properties
`TreeWalkerDataContext` | A `TreeWalkerDataContext` that unites the
`TreeWalkerForTest` with its associated `CancellationToken` and
`Registrar`

It's worth noting that `IActionContext.Registrar` and
`IActionControl.TreeWalkerDataContext.Registrar` refer to the same
object. The overlap could be removed, but then Actions code that has no
relationship to `TreeWalker` objects would be accessing the data through
the `TreeViewDataContext`, which looked wrong. This is easy to change if
we decide to.

There are 2 implementations of `IActionContext`
- `DefaultActionContext` connects everything to the singleton objects
that the previous code used. This object triggers no cleanup on its
`Dispose` method, just like the existing code.
- `ScopedActionContext` connects everything to scoped-lifetime objects.
This object triggers cleanup when it's disposed to (hopefully) prevent
memory leaks

Methodology:
1) Find all places where we use a class's `GetDefaultInstance` method
2) Plumb these through using members from `IActionContext`, working up
the stack to Actions
3) At the Action level, split the methods into 2 methods--an internal
method that takes an `IActionContext` parameter, and the old public
method that calls the internal method, passing in the
`DefaultActionContext`. The intended goal is that AIWin can call the
existing API and keep its existing behavior
4) In the `Automation` namespace, create a `ScopedActionContext` for the
duration of the scan, then use it to call the internal version of the
existing API.

Hopefully, when all of the pieces are brought together, we'll end up
with AIWin working as it currently does, while the Automation interface
can maintain distinct sets of data for each scan. We also now have the
plumbing to get a caller-provided `CancellationToken` down to the points
in the code where we check for cancellation.

While working on this, I ran into problems with
`SelectAction.SetCandidateElementFromProcessId` and
`SelectActcion.SetCandidateElementFromHandle`, which have corresponding
methods in the `A11yAutomation` class. These problems weren't
insurmountable (we just needed to store the `Registrar` in the
`SelectionAction`), but since AIWin never calls these methods, it was
cleaner to remove them than to fix them.

##### Motivation

Part of async feature

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue:
[#1996306](https://mseng.visualstudio.com/1ES/_workitems/edit/1996306)