feat(async-api): Multiple scan roots are always enabled (#763)
#### Details

This is a BREAKING change to the CLI. Scripts that called the CLI with
the --aremultiplescanrootsenabled parameter will reject the arguments.
To be fair, though, we know of nobody who uses this parameter. The only
known case of scanning multiple top-level windows goes directly to the
Automation API and does not use the CLI.

When we added multiple scan roots, we jumped through some hoops to
ensure it would be non-breaking. The 2.0 Automation API is a breaking
change, so we've decided to simplify multiple scan roots. They're always
enabled and currently have no means of disabling them. If requested,
it's easy to add at a later time.

One other change in behavior to call out: For anyone who used the
Automation API and specified multiple root window support, scans would
generate an a11ytestfile even if only one top-level window was scanned
and no errors were found. This changes this. When a single top-level
window is scanned, the a11ytest file will be generated only if errors
exist. When multiple top-level windows are scanned, an a11ytest file
will be generated per top-level window, regardless of error status.

##### Motivation

- Part of async feature work
- Simplify usage
- Steer people towards what we consider to be the long-term scenario.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
There is no way to say "Only scan one top-level window if multiple
top-level windows exist in the process". This is by design, as the order
in which top-level windows is deterministic but hard to concisely
describe. We punted support of this before, and are not adding it here.

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: