chore: Enable CodeQL analysis (#765)
#### Details

CodeQL analysis should be available via injection in default build
pipelines. This PR sets the flag that is supposed to enable it. An
intermediate commit included explicit tasks, which are supposedly not
necessary in the default branch. We'll find out after merging this and
triggering a build.

##### Motivation

New security requirement

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
This does not yet enable TSA uploads. That may occur at a later time

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: