feat(async-api): Refactor A11yAutomation class (#772)
#### Details

In #767, we agreed to revisit refactoring the `A11yAutomation` class to
avoid the possibility of the `DesktopDataContext` referencing a
different `A11yAutomation` object. This PR does the following:
1. Removes the unused `GetFocusedElement` method (I confirmed it's not
used by AIWin)
2. Makes the `GetAppElement` and `NormalizedElementFromPoint` static to
avoid the chance their `DesktopDataContext` parameters refer to a
different `A11yAutomation` object
3. Adds necessary internal checks to guard against a null
`DesktopDataContext`
4. Removes the now-superfluous `EnsureContextConsistency` method
5. Updates the calling code
6. Tightens function visibility on the `A11yAutomation` object.

##### Motivation

<!-- This can be as simple as "addresses issue #123" -->

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
I realized that the `TargetElementLocator` in the Automation assembly
has a direct dependency on `A11yAutomation.ElementsFromProcessId`. That
sort of violates our intended encapsulation of Automation doing
everything through Actions, but the fix there is to refactor the
`TargetElementLocator` class, which feels like a separate PR

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] Addresses an existing issue: #0000