chore: Fix malformed rule description (#775)
#### Details

We noticed that the description for the
`LocalizedLandmarkTypeExcludesPrivateUnicodeCharacters` was malformed.
It read:

The [StringProperty not set] property must not contain any characters in
the private Unicode range.

It should read:

The LocalizedControlType property must not contain any characters in the
private Unicode range.

Easy fix, just add the description to the property and rebuild
RulesDescription.md

##### Motivation

Docs should be correct

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: