feat(async-api): Restore old signature (#776)
#### Details

Changes in #767 removed a parameter and from the public constructor of
`DesktopElementAncestry` and its associated plumbing, since the field
was never used in Axe.Windows. Unfortunately, AIWIn relies on this
functionality
[here](https://github.com/microsoft/accessibility-insights-windows/blob/main/src/AccessibilityInsights.SharedUx/Controls/TestTabs/TabStopControl.xaml.cs#L419).
This PR restores the property with a tiny refactor to (hopefully) make
the code clearer.

I've confirmed that AIWin builds correctly with this change

##### Motivation

Part of async feature work

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: