feat(localization): Add script to check resource strings for comments (#787)
#### Details

As the resource strings in the core, rules, and ruleselection projects
will be localized soon, it is important that we include any necessary
translator comments along with those resources. We have gone through and
added comments in previous PRs, but this PR adds a script to our PR
builds to verify that we continue to have translator notes for all
resource strings in those projects.

##### Motivation

Localization feature

##### Context

The script does not include the resource files for projects that are not
localized right now, since there are quite a lot of strings in those
projects that do not have comments yet. If we ever decide to localize
those projects, it will be simple to remove them from the exclude list
in this script.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue:
[2002611](https://dev.azure.com/mseng/1ES/_workitems/edit/2002611)

Co-authored-by: Bill Dengler <codeofdusk@gmail.com>