chore: Enable analyzer flags to catch unneeded "this." prefixes (#796)
#### Details

Enable analyzer rules to call out places where `this.` is used
unnecessarily

##### Motivation

The code was originally written by folks who used `this.` _everywhere_
because that was the pattern in the team that they came from. Over time,
that pattern has fallen by the wayside and now the code is very
inconsistent. This PR is part of a series of PRs intended to increase
the code consistency by removing unnecessary `this`. usage wherever
possible.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: