chore: Use PascalCase for method names (#809)
#### Details

I recently noticed some camelCase function names when we want to use
PascalCase. I found a way to set this in the .editorconfig file and
fixed the cases that I was able to locate. One case removed the
underscore prefix that was used to mark an internally available
function. I renamed it to reflect the fact that it's non-locking (the
conflicting version is locking and calls the non-locking version).

##### Motivation

Code consistency

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue