chore: add combine-dependabot-prs workflow (#808)
#### Details

Add a new workflow to automatically combine dependabot PRs into a single
PR. Any PR with a failing build or the `pr: do not combine` label will
be excluded from the new combined PR.

This workflow runs automatically every Monday at 1PM UTC (6AM PST), or
it can be run manually from the "actions" tab in the repo.

##### Motivation

Reduce the number of dependabot commits in the repo

##### Context

This action is already used in the
[accessibility-insights-service](https://github.com/microsoft/accessibility-insights-service)
repo on a manual trigger. After a discussion, our team has decided to
experiment with using this action on an automatic trigger in our other
repos.