chore: Update private field names in RulesTest project (#818)
#### Details

Code standards are very inconsistent throughout the code base. This PR
contains the changes needed to make the `RulesTests` project consistent
after locally applying #810. The renames were all done through the IDE,
and in most cases the IDE-suggested name was used. In cases where the
IDE-supplied name seemed odd, I chose a name that seemed appropriate.
There were also several cases where the simplest fix was to make a
member static.

##### Motivation

Improve code consistency and clarity

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
A separate PR will be used to normalize the use of the `readonly`
keyword in this project

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue