chore: Miscellaneous cleanup (#821)
#### Details

A few general cleanup changes:
1. Disable the `dotnet_style_prefer_compound_assignment` rule in
.editorconfig, since it's a C# 8.0 feature and .NET standard 2.0 assumes
C# 7.3 support
2. Remove private setters that aren't needed. Omitting the private
setter means that the property can only be set in the constructor.
Methodology was to remove _all_ private setters, then bring back the
ones that the compiler said we need
3. Rename the `EventListenerBase.disposedValue` property to
`DisposedValue` (it's protected, so the .editorconfig rule didn't catch
it). This updated its usage in all derived classes, as well.
4. Remove some unreferenced fields (called out by the IDE)
5. Remove extra newlines that are floating around
6. Run the IDE's code cleanup agent to tidy up using blocks, adjust
whitespace in lines, etc. This always cuts a little too deep with the
Custom Actions, so some of these changes were manually backed out to
allow the build to succeed

##### Motivation

After this, I think we're .editorconfig clean

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: