chore: Normalize readonly usage in RulesTest project (#820)
#### Details

When we first added the `.editorconfig` file, we disabled the rule that
wants the `readonly` attribute in the `RulesTest` project, just to keep
the commit from getting even larger. This PR is the second pass that
enables the `readonly` rule in `RulesTest`, then updates the files as
needed.

##### Motivation

Code consistency

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue