chore: Provide consistent spacing after license header (#822)
#### Details

This PR ensures that we have a newline between the license header and
the first using line. It was done via a global search & replace across
the entire solution, so it should have captured all remaining files that
didn't have the newline.

I can split this up into per-project PR's if desired, but since the
reviews are so straightforward I started with a single PR

##### Motivation

Code consistency

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: