fix: Add MSAA role name to LegacyIAccessiblePattern.Role (#824)
#### Details

AIWin
[#1454](https://github.com/microsoft/accessibility-insights-windows/issues/1454)
requests that we add the enum name for MSAA role types. This change
provides the mapping, and it will light up in AIWin after this change
gets published and picked up.

We add the `LegacyIAccessibleRoleType` class to map from numeric values
to MSAA role types. We then add this to `A11yProperty` to capture this
on the element. Finally, we need to expose this via the
`A11yPatternProperty`, which leverages the `LegacyIAccessibleRoleType`
class to get the name. In order to avoid confusing quotation marks
around the mapped role, we add
`A11yPatternProperty.OmitQuotesFromString`, which gets set only for
`LegacyIAccessiblePattern.Role`. When writing out the JSON file, we set
the flag to omit null values, so
`A11yPatterProperty.OmitQuotesFromString` only gets written to disk when
it's specified.

Validated by building locally and testing against the pentest site
provided in the AIWin issue.

Two notes:

1. If an `.a11ytest` file is created with the new build, then opened in
a previous version of AIWin, the properties will contain the
LegacyIAccessiblePattern.Role value in quotes. This is because older
versions don't know about the `A11yPatternProperty.OmitQuotesFromString`
property and include the quotes
2. Since AIWin displays the data stored in the `.a11ytest` file,
existing `.a11ytest` files will continue to display just the number
without the enum name
 
##### Motivation

Set things up to address AIWin
[#1454](https://github.com/microsoft/accessibility-insights-windows/issues/1454)

##### Screenshots

Scenario | Before PR | After PR
--- | --- | ---
Live mode (pen test) |
![image](https://user-images.githubusercontent.com/45672944/200637655-90a4a029-093c-4590-be55-b574d50750fe.png)
|
![image](https://user-images.githubusercontent.com/45672944/200638440-83e72f68-293f-4bd5-97a0-23f5c7c272ff.png)
Open old saved file |
![image](https://user-images.githubusercontent.com/45672944/200637885-1cefafb4-21aa-44f4-ae62-3a2c6779b550.png)
|
![image](https://user-images.githubusercontent.com/45672944/200637392-af23c5f8-5012-4abd-a487-5b2993bfd0ee.png)
Open new saved file |
![image](https://user-images.githubusercontent.com/45672944/200638145-fa4a78bb-7508-4d2f-b433-dc3ce4e0d0f8.png)
|
![image](https://user-images.githubusercontent.com/45672944/200637248-34c41862-f9ec-4460-9576-372a2311b3f8.png)


##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] Addresses an existing issue: AIWin
[#1454](https://github.com/microsoft/accessibility-insights-windows/issues/1454)