chore: Tiny code cleanups that got missed before (#830)
#### Details

This PR cleans up 2 things:
1. Renaming `Registrar.sDefaultInstance` to `Registrar.DefaultInstance`
to match naming guidelines
2. Removing an unused parameter from an extension method. The entire
method itself seems to be unused but I left it for now.

##### Motivation

Code consistency

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
I'd love to remove all unused methods, but I'm not sure how to identify
them in an efficient way

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: