chore: Migrate EXE's from .NET Core 3.1 to .NET 6.0 (#833)
#### Details

As called out in this [blog
post](https://devblogs.microsoft.com/dotnet/net-core-3-1-will-reach-end-of-support-on-december-13-2022/),
.NET Core 3.1 ends LTS on 13 December 2022, which is coming up quickly.
Some of our .NET dependencies are dropping support for .NET Core 3.1, so
in order to stay on current binaries, we are shifting our EXE's to .NET
6.0. This includes the CLI and all test projects. The libraries in the
NuGet package will continue to build using .NET Standard 2.0, so they'll
continue to work in either .NET or .NET Framework applications. They
will also continue to be Windows-specific, which makes sense since UIA
is a Windows-specific technology.

Validation included:
1. Run changes through the signed build pipeline. Artifacts are
[here](https://dev.azure.com/mseng/1ES/_build/results?buildId=18716735&view=artifacts&pathAsName=false&type=publishedArtifacts)
2. Install and run the signed CLI (using the MSI installer)
3. Using a local package store, update AIWin to use the signed NuGet
package and make sure that it still works as expected

##### Motivation

Avoid the situation where we depend on libraries that are out of support
or untested in our install environment.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
This and #834 both involve the WXS file. Some merge resolution will be
needed, no matter what order we merge.

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: