chore: Add wxs validation/fix gaps (#834)
#### Details

In
https://github.com/microsoft/accessibility-insights-windows/pull/1388,
we added a WXS validation check to the AIWin MSI definition to identify
gaps that are introduced by package updates (mainly driven by
dependabot). It seemed like a good idea to have a similar check for
Axe.Windows. When running the test, it identified gaps that had been
introduced by package updates, so those are includes in this PR.

Note that since the MSI build is limited release builds, so are the
MsiFileTests.

##### Motivation

Ensure that the CLI always has the expected set of files.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: