fix: Signed build is failing WXS validation (#840)
#### Details

The signed release build is including a localization file that is not
included in the other builds. Since the CLI is not yet localized, we are
not including the localized files and should add this file to the list
of files that we intentionally exclude from the MSI package.

##### Motivation

Unblock signed builds

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: