chore: Fix CLI validation script (#842)
#### Details

When #833 changed the CLI to target .NET 6.0, it changed the path of the
locally-built CLI, which (in turn) broke the validation script. This
just fixes the validation script and makes the displayed instructions a
little bit clearer.

##### Motivation

Validation scripts should work

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: