chore: Fix gaps in WxsValidationTests (#845)
#### Details

This is the analog to [AIWin
1483](https://github.com/microsoft/accessibility-insights-windows/pull/1483).

This fixes some issues with the MSI checker:
1. Due to a logic error, files were returned _only_ if a set of excluded
files was provided. This empty set was not caught as a problem. These
have both been fixed.
2. We weren't passing the correct component names in 1, so we were
returning empty sets of files. This empty set was also not caught as a
problem. This has been fixed.
3. As written, the checker assumed that everything in a component comes
from the same build folder, which is OK in this project. For consistency
with AIWIn, though, we ignore the source path of the file and validate
the _set_ of files in the folder.

There was also 1 place in the WXS file where we were using a component
name that was originally derived from a deprecated version of .NET Core.
This component and its containing group have been renamed for clarity.
This has absolutely no impact on the files included in the MSI package.
Happy to move that to a different PR if desired.

##### Motivation

Make test detect errors (which is why it's there)

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->
It might also make sense to rename `ProductComponents` to
`ProductComponentGroup` in the WXS file. It's easy to change this at any
time if we so desire.

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [n/a] Addresses an existing issue: